# Compile regexp in detokenizer