@ZJaume 2023-09-14
I don't want to be picky, but does that big.txt contain tokenized sentences? Performance may be different if input is not tokenized?